Remove one-page-sections plugin.
This is a small-enough code commit, but it's actually a pretty
substantial change, so I'll try to get it all down.

Got rid of my One Page Sections plugin.
It wasn't bad code or anything, but the idea was just fucked,
unfortunately.  Structurally, it was cool, but hand-writing HTML
into the stupid WYSIWYG text editor for all posts (and therefore
fighting WordPress' default rules) was just a bad way to maintain
content for non-developers.  And for developers as well -- who
are we kidding.

- Deactivated one_page_sections plugin in the plugins interface
- Removed the custom field `acf_checkboxes` that was looking for
which page was meant to be the page used as the One Page Sections
page
- Created a new page template -- `page-home.php` -- which just does
the home page.  That's all it does.

That's it.

Worth noting that the HTML is still sucky.  Every page has two
`h1` elements on it, there's no 'skip' link, and the homepage
breaks the HTML that all other pages use. :/

Shouldn't be impossible to fix, but not right now.