Fixed stupid grunt problem.
For some reason, the last time I modified the `Gruntfile.js`, I
left it in a state where the current CSS would be compiled to a
minified version, and then vendor prefixes would be added to the
non-minified verison.  So neither was really working and I don't
know when that change went in but it doesn't sound that awesome.