Make "back to top" button slightly bigger
Somehow the back arrow on the overlay ended up slightly larger
than the "back to top" button, which I positioned in a similar
part of the screen to take advantage of the fact that they are
the same element, the same font size, only one of them is rotated.

Now the design works.