smaller titles and metas on phones
Basically, we want the titles and meta text underneath to look
better on phones, so we have to make the letters smaller on really
tiny screens but only for projects (which have an archive link in
the header (ie, "::projects") and have an accompanying image.

On single pages, we have less elements on the page, so we can keep
the larger titles.